feat(metrics): Add basic metrics for consensus and gossip (#220)
* feat(metrics): Add basic metrics for consensus and associated configuration

* Add `malachite_consensus_time_per_block` metric

* Cleanup

* Track remaining metrics

* Export libp2p and GossipSub metrics using a prefixed registry

- Gossip consensus metrics are prefixed with `malachite_gossip_consensus_`
- Gossip mempool metrics are prefixed with `malachite_gossip_mempool_`

* Pass metrics registry around instead of relying on global one

* Formatting

* Fix prefix

* Fix span

* Add `metrics.enabled` config option

* Cleanup

* Add FIXME

+++ code/crates/common/src/transaction.rs
@@ -17,8 +17,8 @@ impl Transaction {
     }
 
     /// Size of this transaction in bytes
-    pub fn size_bytes(&self) -> u64 {
-        self.0.len() as u64
+    pub fn size_bytes(&self) -> usize {
+        self.0.len()
     }
 }
